rfc5321::client

Enum ClientError

pub enum ClientError {

    Utf8(FromUtf8Error),
    MalformedResponseLine(String),
    ResponseTooLong,
    NotConnected,
    Rejected(Response),
    RejectedBatch(Vec<Response>),
    InvalidDnsName(String),
    FailedToBuildConnector {
        error: String,
    },
    TimeOutResponse {
        command: Option<Command>,
        duration: Duration,
    },
    TimeOutRequest {
        commands: Vec<Command>,
        duration: Duration,
    },
    ReadError {
        command: Option<Command>,
        error: String,
        partial: String,
    },
    FlushError {
        error: String,
    },
    WriteError {
        commands: Vec<Command>,
        error: String,
    },
    TimeOutData,
    SslErrorStack(ErrorStack),
    NoUsableDaneTlsa {
        hostname: String,
        tlsa: Vec<TLSA>,
    },

}

Variants

Utf8(FromUtf8Error)

MalformedResponseLine(String)

ResponseTooLong

NotConnected

Rejected(Response)

RejectedBatch(Vec<Response>)

InvalidDnsName(String)

FailedToBuildConnector

Fields

error: String

TimeOutResponse

Fields

command: Option<Command>

duration: Duration

TimeOutRequest

Fields

commands: Vec<Command>

duration: Duration

ReadError

Fields

command: Option<Command>

error: String

partial: String

FlushError

Fields

error: String

WriteError

Fields

commands: Vec<Command>

error: String

TimeOutData

SslErrorStack(ErrorStack)

NoUsableDaneTlsa

Fields

hostname: String

tlsa: Vec<TLSA>

Implementations

impl ClientError

pub fn command(&self) -> Option<String>

Returns the command(s) string suitable for passing into a Response

pub fn was_due_to_message(&self) -> bool

If the error contents were likely caused by something about the mostly recently attempted message, rather than a transport issue, or a carry-over from a prior message (eg: previous message was rejected and destination chose to drop the connection, which we detect later in RSET on the next message), then we return true. The expectation is that the caller will transiently fail the message for later retry. If we return false then the caller might decide to try that same message again more immediately on a separate connection