rfc5321

Struct TlsOptions

pub struct TlsOptions {
    pub insecure: bool,
    pub alt_name: Option<String>,
    pub dane_tlsa: Vec<TLSA>,
    pub prefer_openssl: bool,
    pub certificate_from_pem: Option<Arc<Box<[u8]>>>,
    pub private_key_from_pem: Option<Arc<Box<[u8]>>>,
    pub openssl_cipher_list: Option<String>,
    pub openssl_cipher_suites: Option<String>,
    pub openssl_options: Option<SslOptions>,
    pub rustls_cipher_suites: Vec<SupportedCipherSuite>,
}

Fields
insecure: boolalt_name: Option<String>dane_tlsa: Vec<TLSA>prefer_openssl: boolcertificate_from_pem: Option<Arc<Box<[u8]>>>private_key_from_pem: Option<Arc<Box<[u8]>>>openssl_cipher_list: Option<String>openssl_cipher_suites: Option<String>openssl_options: Option<SslOptions>rustls_cipher_suites: Vec<SupportedCipherSuite>
Implementations

impl TlsOptions

pub async fn build_tls_connector(&self) -> Result<TlsConnector, Error>

Produce a TlsConnector for this set of TlsOptions. We need to employ a cache around the verifier as loading the system certificate store can be a non-trivial operation and not be something we want to do repeatedly in a hot code path. The cache does unfortunately complicate some of the internals here.

pub fn build_openssl_connector( &self, hostname: &str, ) -> Result<ConnectConfiguration, OpensslConnectorError>

Build an OpenSSL connector configuration for use with DANE TLSA or when OpenSSL-specific features are needed.