Trait rfc5321::traits::AsyncReadAndWrite

pub trait AsyncReadAndWrite: AsyncRead + AsyncWrite + Debug + Unpin + Send {
    // Provided method
    fn try_dup(&self) -> Option<TcpStream> { ... }
}

Provided Methods

fn try_dup(&self) -> Option<TcpStream>

Optionally clone a TcpStream that represents the same underlying stream as this one. This only has an impl that returns Some for TcpStream. It is present to facilitate a workaround for some awkwardness in the SslStream implementation for the failed-handshake case.